requirements.txt
open_clip
torch
Pillow
opencv-python
openai
pydub
moviepy